requirement.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
idna==2.10
pdf2image==1.14.0
pdfminer.six==20200726
Pillow==7.2.0
psycopg2==2.8.6
pycparser==2.20
pytesseract==0.3.6
requests==2.24.0
six==1.15.0
sortedcontainers==2.2.2
soupsieve==2.0.1
urllib3==1.25.10
